pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "z"
version = "0.6.0"
description = "Starter z package"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "z contributors" }]
dependencies = [
  "pycryptodome>=3.20.0",
  "Pillow>=10.0.0",
]

[project.scripts]
zi = "z.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["z*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
